modularize lstm ops
