+++ tfjs-core/src/index.ts
@@ -64,7 +64,6 @@ export {GradSaveFunc, NamedTensorMap, TensorContainer, TensorContainerArray, Ten
 export {BackendValues, DataType, DataTypeMap, DataValues, NumericDataType, PixelData, Rank, RecursiveArray, ShapeMap, sumOutType, TensorLike, TypedArray, upcastType} from './types';
 
 export * from './ops/ops';
-export {LSTMCellFunc} from './ops/lstm';
 export {Reduction} from './ops/loss_ops';
 
 export * from './train';
