+++ tfjs-core/src/ops/multi_rnn_cell.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2018 Google Inc. All Rights Reserved.
+ * Copyright 2020 Google Inc. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,8 +14,7 @@
  * limitations under the License.
  * =============================================================================
  */
-
-import {Scalar, Tensor1D, Tensor2D} from '../tensor';
+import {Tensor2D} from '../tensor';
 import {convertToTensor, convertToTensorArray} from '../tensor_util_env';
 import {TensorLike} from '../types';
 import {op} from './operation';
@@ -66,53 +65,4 @@ function multiRNNCell_(
   }
   return [newC, newH];
 }
-
-/**
- * Computes the next state and output of a BasicLSTMCell.
- *
- * Returns `[newC, newH]`.
- *
- * Derived from tf.contrib.rnn.BasicLSTMCell.
- *
- * @param forgetBias Forget bias for the cell.
- * @param lstmKernel The weights for the cell.
- * @param lstmBias The bias for the cell.
- * @param data The input to the cell.
- * @param c Previous cell state.
- * @param h Previous cell output.
- */
-/** @doc {heading: 'Operations', subheading: 'RNN'} */
-function basicLSTMCell_(
-    forgetBias: Scalar|TensorLike, lstmKernel: Tensor2D|TensorLike,
-    lstmBias: Tensor1D|TensorLike, data: Tensor2D|TensorLike,
-    c: Tensor2D|TensorLike, h: Tensor2D|TensorLike): [Tensor2D, Tensor2D] {
-  const $forgetBias =
-      convertToTensor(forgetBias, 'forgetBias', 'basicLSTMCell');
-  const $lstmKernel =
-      convertToTensor(lstmKernel, 'lstmKernel', 'basicLSTMCell');
-  const $lstmBias = convertToTensor(lstmBias, 'lstmBias', 'basicLSTMCell');
-  const $data = convertToTensor(data, 'data', 'basicLSTMCell');
-  const $c = convertToTensor(c, 'c', 'basicLSTMCell');
-  const $h = convertToTensor(h, 'h', 'basicLSTMCell');
-
-  const combined = $data.concat($h, 1);
-  const weighted = combined.matMul($lstmKernel);
-  const res: Tensor2D = weighted.add($lstmBias);
-
-  // i = input_gate, j = new_input, f = forget_gate, o = output_gate
-  const batchSize = res.shape[0];
-  const sliceCols = res.shape[1] / 4;
-  const sliceSize: [number, number] = [batchSize, sliceCols];
-  const i = res.slice([0, 0], sliceSize);
-  const j = res.slice([0, sliceCols], sliceSize);
-  const f = res.slice([0, sliceCols * 2], sliceSize);
-  const o = res.slice([0, sliceCols * 3], sliceSize);
-
-  const newC: Tensor2D = i.sigmoid().mul(j.tanh()).add(
-      $c.mul($forgetBias.add(f).sigmoid() as Tensor2D));
-  const newH: Tensor2D = newC.tanh().mul(o.sigmoid());
-  return [newC, newH];
-}
-
-export const basicLSTMCell = op({basicLSTMCell_});
 export const multiRNNCell = op({multiRNNCell_});
