+++ tfjs-core/src/ops/ops.ts
@@ -21,6 +21,7 @@ export {addN} from './add_n';
 export {atan2} from './atan2';
 export {avgPool} from './avg_pool';
 export {avgPool3d} from './avg_pool_3d';
+export {basicLSTMCell} from './basic_lstm_cell';
 export {batchToSpaceND} from './batch_to_space_nd';
 export {batchNorm} from './batchnorm';
 export {batchNorm2d} from './batchnorm2d';
@@ -72,6 +73,7 @@ export {maximum} from './maximum';
 export {minimum} from './minimum';
 export {mod} from './mod';
 export {mul} from './mul';
+export {LSTMCellFunc, multiRNNCell} from './multi_rnn_cell';
 export {multinomial} from './multinomial';
 export {notEqual} from './not_equal';
 export {oneHot} from './one_hot';
@@ -116,7 +118,6 @@ export * from './transpose';
 export * from './softmax';
 export * from './norm';
 export * from './segment_ops';
-export * from './lstm';
 export * from './moving_average';
 export * from './strided_slice';
 export * from './topk';
