+++ tfjs-core/src/tests.ts
@@ -46,6 +46,7 @@ import './ops/avg_pool_3d_test';
 import './ops/avg_pool_test';
 import './ops/axis_util_test';
 import './ops/band_part_test';
+import './ops/basic_lstm_cell_test';
 import './ops/batch_to_space_nd_test';
 import './ops/batchnorm_test';
 import './ops/binary_ops_test';
@@ -91,12 +92,12 @@ import './ops/logical_not_test';
 import './ops/logical_or_test';
 import './ops/logical_xor_test';
 import './ops/loss_ops_test';
-import './ops/lstm_test';
 import './ops/mat_mul_test';
 import './ops/max_pool_3d_test';
 import './ops/max_pool_test';
 import './ops/max_pool_with_argmax_test';
 import './ops/moving_average_test';
+import './ops/multi_rnn_cell_test';
 import './ops/multinomial_test';
 import './ops/non_max_suppression_async_test';
 import './ops/non_max_suppression_test';
